Theoretically reversed a Menu Item Option into a service
I say reversed, because I started at the database and worked up through
a repo, api and then service.  It isn't attached to any UI, but I still
have to consider the best way to do that - especially as I need to know
the MenuItemId before I can really save the damn thing.